Update easyim: seperate infgraph to more files.
