+++ EaSyIM/src/infgraph.h
@@ -2,11 +2,6 @@
 // Created by cuonghn on 26/5/17.
 //
 
-//#ifndef IM_INFGRAPH_H
-//#define IM_INFGRAPH_H
-//
-//#endif //IM_INFGRAPH_H
-
 class InfGraph : public Graph{
 private:
     vector<bool > visit;
@@ -45,6 +40,8 @@ public:
     deque<int> q;
     sfmt_t sfmtSeed;
     set<int> seedSet;
+#include "assignscore.h"
+#include "activatenodes.h"
     void BuildSeedSetGreedy(int k, int l)
     {
         seedSet.clear();
@@ -57,100 +54,6 @@ public:
             int id = t - cur_score.begin();
             seedSet.insert(id);
             UpdateActivatedNodes(id);
-        }
-    }
-
-    void EasyIMAssignScore(int max_length)
-    {
-        init_score();
-        if (influModel == IC)
-        {
-            // for each node
-            for (int i = 0; i < n; i++)
-            {
-                // for each outgoing edge from node i
-                if (visit[i]) continue;
-                for (int j = 0; j < (int)g[i].size(); j++)
-                {
-                    int v = g[i][j];
-                    if (!visit[v])
-                    {
-                        new_score[i] += 0.1 * (1 + cur_score[v]);
-                    }
-                }
-            }
-            // update current score after each step
-            cur_score = new_score;
-        } else
-        for (int step=0; step < max_length; step++)
-        {
-            // for each node
-            for (int i = 0; i < n; i++)
-            {
-                // for each outgoing edge from node i
-                if (visit[i]) continue;
-                for (int j = 0; j < (int)g[i].size(); j++)
-                {
-                    int v = g[i][j];
-                    if (!visit[v])
-                    {
-                        new_score[i] += prob[i][j] * (1 + cur_score[v]);
-                    }
-                }
-            }
-            // update current score
-            cur_score = new_score;
-        }
-    }
-
-    void UpdateActivatedNodes(int seed_node)
-    {
-        q.clear();
-        q.push_back(seed_node);
-        visit[seed_node] = true;
-        // update visited nodes start from seed node
-        if (influModel == WC)
-        {
-            while (!q.empty())
-            {
-                int expand = q.front();
-                q.pop_front();
-                int i = expand;
-                for (int j = 0; j < (int) g[i].size(); j++)
-                {
-                    int v = g[i][j];
-                    double randDouble = sfmt_genrand_real1(&sfmtSeed);
-                    if (randDouble > prob[i][j])
-                        continue;
-                    if (visit[v])
-                        continue;
-                    if (!visit[v])
-                    {
-                        visit[v] = true;
-                    }
-                    q.push_back(v);
-                }
-
-            }
-        } else if (influModel == IC)
-        {
-            int expand = q.front();
-            q.pop_front();
-            int i = expand;
-            for (int j = 0; j < (int) g[i].size(); j++)
-            {
-                int v = g[i][j];
-                double randDouble = sfmt_genrand_real1(&sfmtSeed);
-                if (randDouble > 0.1)
-                    continue;
-                if (visit[v])
-                    continue;
-                if (!visit[v])
-                {
-                    visit[v] = true;
-                }
-                q.push_back(v);
-            }
         }
     }
 };
